src/main/java/testinglab/ArrayReversor.java
package testinglab;

import java.util.ArrayList;
import java.util.List;

public class ArrayReversor {
    private ArrayFlattenerService flattenerService;

    public ArrayReversor(ArrayFlattenerService flattenerService) {
        this.flattenerService = flattenerService;
    }

    public int[] reverseArray(int[][] input) {
        if (input == null) {
            return new int[0]; // or handle as needed
        }
        
        int[] flattenedArray = flattenerService.flattenArray(input);
        int[] reversedArray = new int[flattenedArray.length];

        for (int i = 0; i < flattenedArray.length; i++) {
            reversedArray[i] = flattenedArray[flattenedArray.length - 1 - i];
        }
        return reversedArray;
    }
}
